move server-only stuff to ~/server
