+++ pages/api/pages.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import nc from "next-connect";
 
-import { createPage } from "@server/handlers/pageHandlers";
+import { createPage } from "~/server/handlers/pageHandlers";
 
 const handler = nc<NextApiRequest, NextApiResponse>().post(createPage);
