+++ pages/api/pages/[pageId].ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import nc from "next-connect";
 
-import { deletePage, getPage, updatePage } from "@server/handlers/pageHandlers";
+import { deletePage, getPage, updatePage } from "~/server/handlers/pageHandlers";
 
 const handler = nc<NextApiRequest, NextApiResponse>()
     .get(getPage)
