+++ pages/api/pages/[pageId]/comments.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import nc from "next-connect";
 
-import { createPageComment } from "@server/handlers/pageHandlers";
+import { createPageComment } from "~/server/handlers/pageHandlers";
 
 const hander = nc<NextApiRequest, NextApiResponse>().post(createPageComment);
