+++ pages/api/pages/[pageId]/comments/[commentId].ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import nc from "next-connect";
 
-import { deletePageComment, updatePageComment } from "@server/handlers/pageHandlers";
+import { deletePageComment, updatePageComment } from "~/server/handlers/pageHandlers";
 
 const handler = nc<NextApiRequest, NextApiResponse>()
     .post(updatePageComment)
