+++ pages/api/sites.ts
@@ -1,8 +1,8 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import nc from "next-connect";
 
-import { createSite } from "@server/handlers/siteHandlers";
-import { removeCreateSiteRequestProps } from "@server/middlewares/removeProps";
+import { createSite } from "~/server/handlers/siteHandlers";
+import { removeCreateSiteRequestProps } from "~/server/middlewares/removeProps";
 
 const handler = nc<NextApiRequest, NextApiResponse>().post(
     removeCreateSiteRequestProps,
