+++ pages/api/sites/[siteId].ts
@@ -1,8 +1,8 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import nc from "next-connect";
 
-import { deleteSite, getSite, updateSite } from "@server/handlers/siteHandlers";
-import { removeUpdateSiteRequestProps } from "@server/middlewares/removeProps";
+import { deleteSite, getSite, updateSite } from "~/server/handlers/siteHandlers";
+import { removeUpdateSiteRequestProps } from "~/server/middlewares/removeProps";
 
 const handler = nc<NextApiRequest, NextApiResponse>()
     .get(getSite)
