+++ pages/api/users/[uid].ts
@@ -1,8 +1,8 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import nc from "next-connect";
 
-import { deleteUser, getUser, updateUser } from "@server/handlers/userHandlers";
-import { validateUidWithJWT } from "@server/middlewares/validateRequests";
+import { deleteUser, getUser, updateUser } from "~/server/handlers/userHandlers";
+import { validateUidWithJWT } from "~/server/middlewares/validateRequests";
 
 const handler = nc<NextApiRequest, NextApiResponse>()
     .get(getUser)
