+++ pages/api/users/[uid]/sites.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import nc from "next-connect";
 
-import { listUserSites } from "@server/handlers/userHandlers";
+import { listUserSites } from "~/server/handlers/userHandlers";
 
 const handler = nc<NextApiRequest, NextApiResponse>({
     onError: (err, _, res) => {
