+++ server/middlewares/removeProps/index.ts
@@ -1,8 +1,8 @@
 /**
  * This module provides middlewares to remove ALL redudant properties of an object.
  */
-import { KeyNameSet } from "@server/types";
-import { INextApiMiddleware } from "@server/types/nextApi.type";
+import { KeyNameSet } from "~/server/types";
+import { INextApiMiddleware } from "~/server/types/nextApi.type";
 
 export function _createRemovePropsMiddleware<T>(allowed: KeyNameSet<T>) {
     const allowedNames = allowed as Set<string>;
