+++ server/types/page.type.ts
