+++ tsconfig.json
@@ -15,10 +15,7 @@
     "isolatedModules": true,
     "jsx": "preserve",
     "baseUrl": "./",
-    "paths": {
-      "~/*": ["./*"],
-      "@server/*": ["./lib/server/*"]
-    }
+    "paths": { "~/*": ["./*"] }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
   "exclude": ["node_modules"]
